# target/classes/openapi.yml
openapi: 3.0.1
info:
  title: Cycle-based Translation Inference Service API
  description: "This API allows to obtain translations from a source language into\
    \ a target language by exploring the connections in a graph of dictionaries. This\
    \ technique was proposed by [1] in 2006. The idea was exploting the properties\
    \ of the Apertium RDF Graph by using cycles to identify potential targets that\
    \ may be a translation of a given word. [1] Villegas, M., Melero, M., Bel, N.,\
    \ & Gracia, J. (2016, May). Leveraging RDF graphs for crossing multiple bilingual\
    \ dictionaries. In Proceedings of the Tenth International Conference on Language\
    \ Resources and Evaluation (LREC'16) (pp. 868-876)"
  version: 1.0.0-oas3
  x-port: 8080
paths:
  /translation/{source}-{target}:
    get:
      tags:
      - translation
      description: "Given source and target languages, returns a whole inferred bilingual\
        \ dictionary. Results can be optionally saved as tsv. This might take a long\
        \ time to run"
      operationId: inferredBilingualDict
      parameters:
      - name: source
        in: path
        required: true
        schema:
          type: string
        example: en
      - name: target
        in: path
        required: true
        schema:
          type: string
        example: fr
      - name: format
        in: query
        required: false
        schema:
          type: string
        example: tsv
      responses:
        "200":
          description: An array of CycleDensity instances
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CycleDensity'
  /translation/{source}-{target}/{entry}:
    get:
      tags:
      - translation
      description: "Given an enry, and source and target languages, returns a list\
        \ of translations for the given entry in the source language (in the form\
        \ of CycleDensity objects)"
      operationId: translationsFromEntry
      parameters:
      - name: source
        in: path
        required: true
        schema:
          type: string
        example: en
      - name: target
        in: path
        required: true
        schema:
          type: string
        example: fr
      - name: entry
        in: path
        required: true
        schema:
          type: string
        example: dog
      responses:
        "200":
          description: An array of CycleDensity instances
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CycleDensity'
  /translation/test/{source}-{target}:
    get:
      tags:
      - translation
      description: A test endpoint for inferredBilingualDict to infer only the first
        four entries in the bilingual dictionary
      operationId: inferredBilingualDictTest
      parameters:
      - name: source
        in: path
        required: true
        schema:
          type: string
        example: en
      - name: target
        in: path
        required: true
        schema:
          type: string
        example: fr
      - name: format
        in: query
        required: false
        schema:
          type: string
        example: tsv
      responses:
        "200":
          description: An array of four CycleDensity instances
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CycleDensity'
components:
  schemas:
    CycleDensity:
      type: object
      properties:
        target:
          type: string
        score:
          type: number
          format: double
        cycle:
          type: array
          items:
            type: string